lgraph-swarm/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.3.0",
    "langchain[openai]>=1.2.6",
    "langgraph>=1.0.6",
    "python-dotenv>=1.2.1",
    "typing-extensions>=4.15.0",
]
